rename api for consistency and add github actions
