+++ .github/pull_request_template.md
@@ -1,0 +1,26 @@
+Github Issue: #NUM _(replace NUM with Github issue number)_
+
+### Description
+
+_Summarize the ticket here_
+
+### Overview of work done
+
+_Summarize the work you did_
+
+### Overview of verification done
+
+_Summarize the testing and verification you've done. This includes unit tests or testing with specific data_
+
+### Overview of integration done
+
+_Explain how this change was integration tested. Provide screenshots or logs if appropriate. An example of this would be a local Harmony deployment._
+
+## PR checklist:
+
+* [ ] Linted
+* [ ] Updated unit tests
+* [ ] Updated changelog
+* [ ] Integration testing
+
+_See [Pull Request Review Checklist](../CONTRIBUTING.md#reviewing) for pointers on reviewing this pull request_
