+++ hydrocron_api/controllers/subset.py
@@ -8,7 +8,7 @@ import time
 from datetime import datetime
 from typing import Generator
 from shapely import Polygon, Point
-from hydrocronapi import hydrocron
+from hydrocron_api import hydrocron
 
 
 logger = logging.getLogger()
