+++ hydrocron_api/hydrocron.py
@@ -9,7 +9,7 @@ from typing import Callable
 import boto3
 import connexion
 
-from hydrocronapi.data_access.db import DynamoDataRepository
+from hydrocron_api.data_access.db import DynamoDataRepository
 
 
 class Context(ModuleType):
